feat: Turn invite bubble into modal
