+++ src/page/index.html
@@ -18,7 +18,5 @@
     <script src="/min/vendor.js?{{@config.VERSION}}"></script>
     <script src="/min/runtime.js?{{@config.VERSION}}"></script>
     <script src="/min/app.js?{{@config.VERSION}}"></script>
-
-    #include('partials/template-user-profile.htm')
   </body>
 </html>
