+++ src/page/template/wire-main.htm
@@ -68,6 +68,7 @@
         #include('modal/group-creation.htm')
         #include('modal/user-modal.htm')
         #include('modal/service-modal.htm')
+        #include('modal/invite-modal.htm')
         #include('modal/legal-hold.htm')
         <div class="center-column__overlay"></div>
       </div>
