--- src/style/components/user-profile.less
@@ -1,100 +1,0 @@
-/*
- * Wire
- * Copyright (C) 2018 Wire Swiss GmbH
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see http://www.gnu.org/licenses/.
- *
- */
-
-user-profile {
-  .bubble-content;
-
-  color: var(--background);
-}
-
-// TODO: make generic shift classes
-.user-profile-transition {
-  .hardware-acc;
-
-  height: 100%;
-  transition-duration: @animation-timing-fast;
-  transition-property: transform;
-  // TODO: timing function
-}
-
-.user-profile-default {
-  display: flex;
-  overflow: hidden;
-  height: 100%;
-  flex-direction: column;
-}
-
-// user-profile-default
-.user-profile-header {
-  display: flex;
-  min-height: 80px;
-  flex-direction: column;
-  padding-top: 21px;
-  padding-right: 56px;
-  padding-left: 56px;
-  text-align: center;
-}
-
-.user-profile-details {
-  position: relative;
-  display: flex;
-  flex: 1 1;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.user-profile-details-avatar {
-  .square(@avatar-diameter-xl);
-
-  position: relative;
-}
-
-.user-profile-footer {
-  padding-right: 18px;
-  padding-bottom: 18px;
-  padding-left: 18px;
-
-  > .footer-controls-buttons {
-    display: flex;
-    height: 100%;
-    align-items: center;
-    justify-content: center;
-
-    .button:first-child {
-      margin-right: 16px;
-    }
-  }
-
-  > .footer-connection-request {
-    .label-bold-xs;
-    .flex-center;
-
-    color: #fff;
-    cursor: pointer;
-    text-align: center;
-  }
-
-  > .footer-connection-request {
-    height: @bubble-footer-height - 4px;
-    margin-top: 2px;
-    border-bottom-left-radius: 2px;
-    border-bottom-right-radius: 2px;
-  }
-}
