+++ src/style/foundation/header.less
@@ -17,11 +17,8 @@
  *
  */
 
-@import (less) '../ext/css/normalize.css/normalize.css';
 @import (less) '../../node_modules/@wireapp/antiscroll-2/dist/antiscroll-2.css';
 @import (less) '../../node_modules/simplebar/dist/simplebar.css';
 @import (less) '../../node_modules/highlightjs/styles/github.css';
 @import (less) '../fonts/redacted.css';
 @import (less) '../fonts/zeta-neue.css';
-
-@import '../../ext/less/webapp-module-bubble/webapp-module-bubble';
