+++ src/style/foundation/main.less
@@ -17,6 +17,7 @@
  *
  */
 
+@import 'normalize';
 @import 'themes';
 
 @import '../common/accent-color';
@@ -54,7 +55,6 @@
 @import '../components/group-list';
 @import '../components/user-list';
 @import '../components/user-input';
-@import '../components/user-profile';
 @import '../components/participant-avatar';
 @import '../components/accent-color-picker';
 @import '../components/availability-state';
@@ -124,7 +124,6 @@
 @import '../list/list';
 @import '../list/preferences';
 @import '../list/start-ui';
-@import '../list/invite-bubble';
 @import '../list/takeover';
 @import '../list/temporary-guest';
 
@@ -133,3 +132,4 @@
 @import '../modal/user-modal';
 @import '../modal/service-modal';
 @import '../modal/legal-hold-modal';
+@import '../modal/invite-modal';
