--- src/style/list/invite-bubble.less
@@ -1,121 +1,0 @@
-/*
- * Wire
- * Copyright (C) 2018 Wire Swiss GmbH
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see http://www.gnu.org/licenses/.
- *
- */
-
-@keyframes message-bg-fadein {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 0.32;
-  }
-}
-
-@keyframes message-bg-fadeout {
-  0% {
-    opacity: 0.32;
-  }
-  100% {
-    opacity: 0;
-  }
-}
-
-.invite-bubble {
-  z-index: @z-index-bubble;
-  width: @bubble-width;
-  padding: 0;
-  color: var(--background);
-}
-
-.invite-bubble-inner {
-  background-color: var(--app-bg-secondary);
-  border-radius: 4px;
-}
-
-.invite-bubble-button {
-  .label-bold-xs;
-  .flex-center;
-
-  height: 56px;
-  border-top: 1px solid fade(#000, 16%);
-  cursor: pointer;
-}
-
-.invite-bubble-link {
-  display: flex;
-  height: 100%;
-  flex-direction: column;
-}
-
-.invite-content {
-  flex-grow: 1;
-  padding: 24px;
-}
-
-.invite-header {
-  margin-bottom: 16px;
-  font-size: @font-size-sm;
-  font-weight: @font-weight-bold;
-  line-height: @line-height-lg;
-}
-
-.invite-link-box {
-  position: relative;
-  display: inline-block;
-  width: 100%;
-  height: 144px;
-  border: 1px solid rgba(0, 0, 0, 0.08);
-  margin: 0;
-  border-radius: 2px;
-  font-size: @font-size-xs;
-  line-height: @line-height-sm;
-
-  .fade-wrapper {
-    .full-screen;
-  }
-
-  .bg-animation {
-    animation-duration: @animation-timing-slow;
-    animation-name: message-bg-fadein, message-bg-fadeout;
-    animation-timing-function: @ease-out-quart;
-  }
-
-  .message {
-    .reset-textarea;
-
-    position: relative;
-    overflow: hidden;
-    width: 100%;
-    height: 100%;
-    padding: 8px;
-    background-color: transparent;
-    color: var(--background);
-  }
-}
-
-.invite-link {
-  word-break: break-all;
-}
-
-.invite-footer {
-  .label-bold-xs;
-
-  margin-bottom: 24px;
-  color: var(--foreground);
-  text-align: center;
-}
