+++ src/style/list/start-ui.less
@@ -217,19 +217,6 @@
   margin: 0 16px;
 }
 
-.start-ui-user-bubble {
-  user-profile,
-  .start-ui-service {
-    .bubble-content;
-
-    position: relative;
-    margin: -@bubble-border-radius;
-    background: var(--app-bg-secondary);
-    border-radius: 4px;
-    color: var(--background);
-  }
-}
-
 .start-ui-import-container {
   display: flex;
   flex-grow: 1;
