+++ src/style/modal/invite-modal.less
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2018 Wire Swiss GmbH
+ * Copyright (C) 2019 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,11 +17,17 @@
  *
  */
 
-// https://github.com/wireapp/grunt-npm-bower
+.invite-modal {
+  &__message {
+    height: auto;
+    padding: 8px;
+    color: var(--background);
+  }
 
-module.exports = {
-  options: {
-    cleanTargetDir: true,
-    targetDir: 'src/ext',
-  },
-};
+  &__info {
+    margin: 16px;
+    font-size: 12px;
+    font-weight: 600;
+    text-align: center;
+  }
+}
