+++ src/style/support.less
@@ -17,7 +17,6 @@
  *
  */
 
-@import (less) '../ext/css/normalize.css/normalize.css';
 @import (less) 'fonts/zeta-neue.css';
 
 @import 'common/variables';
